luks2: Take digest length to be compared from expected digest
LUKS1 was limited to a 20 byte digest length. With LUKS2 this limitation
was lifted. However, it seems that for LUKS volumes which were converted
from LUKS1 to LUKS2, the digest length it still 20 bytes. This commit
proposes using the length of the expected digest to determine the digit
length in use. For this purpose, it only compares the first n bytes of
the generated digest with the expected digest.

This fixes compatibility of luks.go with volumes converted from LUKS1.

Fixes: #11

See also: https://github.com/anatol/booster/issues/202